set: add HashSet for types that implement Hash() function
This PR introduces a second Set implementation for types that implement
a Hash function.

There is the Hash type constraint, which can be string-like or int-like.

There is also the HashFunc type constraint, which is satisfied by a
type with Hash() Hash function.

A type that satisfies HashFunc can then be stored in a HashSet.