common: some tweaks for common helper functions
This PR renames TransformUnion to InsertSetFunc, and TransformSlice to
SliceFunc which should be a bit more obvious to understand. Also, the
"Union" methods on the set types are ones that return a newly created set,
and so we are also now more consistent with how those are named.

Also adds a boolean return value to InsertSetFunc, which like the insertion
methods on the set types returns true if the set was modified as a result
of the operation. And adds test cases around being modified or not modified.

Cleans up a few doc strings.