v3: implement Items() method for use with range keyword
Starting with go1.23, Go supports using `range` with custom user types
that satisfy interfaces of the `iter` package. We can make use of this
functionality for iterating over elements of each type of set.

```go
s := set.From([]int{2, 1, 1, 3})
for item := range s.Items() {
  fmt.Println(item)
}
```

This functionality replaces the need for the `ForEach` defined on each
set type, and so that method is removed. This means we need a v3 breaking
change for the release.

Closes #82