remove flaky example JSON marshalling tests (#104)
Because `Set` and `HashSet` are unsorted, the output of marshalling them to JSON
is unstable, and this makes for flaky example tests. Given that we have an
equivalent test for `TreeSet` and we can roughly be assured that Go developers
know how to call the JSON serialization functions, we can just eliminate these.